Some frontend and backend username, email and password verifications added for data consistency and validity.
Still need to have password be updatable.
Also need to update the username and age in the various
locations when user infomation is saved.

Also need to update the validation to realize when
the form information is what the initial information was
without resorting to the Form Reset Cancel button.